* feature/auth:
  mysql 연결 및 활성화
  mockito사용 회원가입 테스트 코드 작성
  사용자 정보 변수 변경(username -> userid) / 사용자 정보를 포함하는 save방식 변경
  자잘한 코드 정리
  HomeController로 통합
  게시글 로그인 사용자 정보 삽입 구현 완료
  댓글 저장 로그인 사용자 정보 삽입 구현 완료
  댓글 로그인 사용자 댓글 담을 준비
  로그인 사용자 정보 담을 준비
  코드정리
  게시물에 넣을 회원이름 변수 변경
  중복 로그인 검사
  코드 정리
  비인증 접근시 거부메시지 출력
  접근 불가 (인가) html 추가

+++ src/main/java/com/sparta/spring_w4_homework/service/BoardService.java
@@ -4,11 +4,12 @@ import com.sparta.spring_w4_homework.model.Board;
 import com.sparta.spring_w4_homework.repository.BoardRepository;
 import com.sparta.spring_w4_homework.requestdto.BoardRequestDto;
 import com.sparta.spring_w4_homework.responsedto.BoardResponseDto;
+
 import lombok.RequiredArgsConstructor;
+
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -16,43 +17,56 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 @Service
 public class BoardService {
-
     private final BoardRepository boardRepository;
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
     //게시글 저장
     @Transactional
-    public String boardsave(BoardRequestDto params){
-        Board board = new Board(params);
+    public String boardsave(BoardRequestDto params, String userid) {
+        if (params.getContentb().equals("")) {
+            return "게시물 내용을 입력해주세요";
+        }
+        Board board = new Board(params, userid);
         boardRepository.save(board);
         return "게시글이 저장되었습니다";
     }
 
     //게시글 전체 조회
-    public List<BoardResponseDto> boardfindAll(){
+    public List<BoardResponseDto> boardfindAll() {
         Sort sort = Sort.by(Sort.Direction.DESC, "modifiedAt");
         List<Board> list = boardRepository.findAll(sort);
         return list.stream().map(BoardResponseDto::new).collect(Collectors.toList());
     }
 
     //게시글 상세 조회
-    public BoardResponseDto boardfindById(Long id){
+    public BoardResponseDto boardfindById(Long id) {
         Board entity = boardRepository.findById(id).orElseThrow(() -> new NullPointerException("존재하지 않는 게시글 입니다."));
         return new BoardResponseDto(entity);
     }
 
     //게시글 수정
     @Transactional
-    public String boardupdate(Long id, BoardRequestDto params){
+    public String boardupdate(Long id, BoardRequestDto params, String userid) {
         Board board = boardRepository.findById(id).orElseThrow(() -> new NullPointerException("존재하지 않는 게시글 입니다."));
-            board.update(params.getTitle(), params.getContentb());
-            return "게시글이 수정 되었습니다.";
+        if (!board.getUserid().equals(userid)) {
+            return "게시글 작성자가 아닙니다.";
+        }
+
+        if (params.getContentb().equals("")) {
+            return "게시물 내용을 입력해주세요";
+        }
+        board.update(params);
+        return "게시글이 수정 되었습니다.";
     }
 
     //게시글 삭제
     @Transactional
-    public String boarddelete(Long id){
+    public String boarddelete(Long id, String userid) {
+        Board board = boardRepository.findById(id).orElseThrow(() -> new NullPointerException("존재하지 않는 게시글 입니다."));
+        if (!board.getUserid().equals(userid)) {
+            return "게시글 작성자가 아닙니다.";
+        }
         boardRepository.deleteById(id);
         return "게시글이 삭제 되었습니다.";
     }
